broke out slot
